main_platform_backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-cors-headers==3.13.0
django-environ==0.7.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.0.0
psycopg2==2.9.3
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.2
whitenoise==6.2.0
